fix: resolve NaN-boxing bugs, type metamethods, and complete test suite
- Fixed NaN-boxing to correctly differentiate INFINITY/NaN from tagged types.
- Fixed 32-bit integer limits and serialization.
- Added missing FILE and SOCKET types to Value representation.
- Fixed OP_GET_TABLE metamethod lookup for userdata and native types.
- All 174 test suite assertions now pass.

+++ src/vm/stdlib_math.cpp
@@ -374,6 +374,6 @@ void registerMathLibrary(VM* vm, TableObject* mathTable) {
     mathTable->set("huge", Value::number(HUGE_VAL));
 
     // Add math.maxinteger and math.mininteger
-    mathTable->set("maxinteger", Value::integer(std::numeric_limits<int64_t>::max()));
-    mathTable->set("mininteger", Value::integer(std::numeric_limits<int64_t>::min()));
+    mathTable->set("maxinteger", Value::integer(std::numeric_limits<int32_t>::max()));
+    mathTable->set("mininteger", Value::integer(std::numeric_limits<int32_t>::min()));
 }
